+++ src/OpenTelemetry.Exporter.OneCollector/Internal/Serialization/LogRecordCommonSchemaJsonSerializer.cs
@@ -26,6 +26,14 @@ internal sealed class LogRecordCommonSchemaJsonSerializer : CommonSchemaJsonSeri
     private static readonly JsonEncodedText SeverityTextProperty = JsonEncodedText.Encode("severityText");
     private static readonly JsonEncodedText SeverityNumberProperty = JsonEncodedText.Encode("severityNumber");
     private static readonly JsonEncodedText BodyProperty = JsonEncodedText.Encode("body");
+    private static readonly JsonEncodedText DistributedTraceExtensionProperty = JsonEncodedText.Encode("dt");
+    private static readonly JsonEncodedText DistributedTraceExtensionTraceIdProperty = JsonEncodedText.Encode("traceId");
+    private static readonly JsonEncodedText DistributedTraceExtensionSpanIdProperty = JsonEncodedText.Encode("spanId");
+    private static readonly JsonEncodedText DistributedTraceExtensionTraceFlagsProperty = JsonEncodedText.Encode("traceFlags");
+    private static readonly JsonEncodedText ExceptionExtensionProperty = JsonEncodedText.Encode("ex");
+    private static readonly JsonEncodedText ExceptionExtensionTypeProperty = JsonEncodedText.Encode("type");
+    private static readonly JsonEncodedText ExceptionExtensionMessageProperty = JsonEncodedText.Encode("msg");
+    private static readonly JsonEncodedText ExceptionExtensionStackTraceProperty = JsonEncodedText.Encode("stack");
 
     private static readonly JsonEncodedText[] LogLevelToSeverityTextMappings = new JsonEncodedText[]
     {
@@ -43,24 +51,34 @@ internal sealed class LogRecordCommonSchemaJsonSerializer : CommonSchemaJsonSeri
         1, 5, 9, 13, 17, 21, 1,
     };
 
-    private static readonly Action<LogRecordScope, Utf8JsonWriter> SerializeScopeItemToJson = (s, w) =>
+    private static readonly Action<LogRecordScope, CommonSchemaJsonSerializationState> SerializeScopeItemToJson = (scope, serializationState) =>
     {
-        foreach (KeyValuePair<string, object?> scopeAttribute in s)
+        var writer = serializationState.Writer;
+
+        foreach (KeyValuePair<string, object?> scopeAttribute in scope)
         {
             if (scopeAttribute.Key == "{OriginalFormat}")
             {
                 return;
             }
 
-            CommonSchemaJsonSerializationHelper.SerializeKeyValueToJson(scopeAttribute.Key, scopeAttribute.Value, w);
+            if (AttributeKeyStartWithExtensionPrefix(scopeAttribute.Key))
+            {
+                serializationState.AddExtensionAttribute(scopeAttribute!);
+                continue;
+            }
+
+            CommonSchemaJsonSerializationHelper.SerializeKeyValueToJson(scopeAttribute.Key, scopeAttribute.Value, writer);
         }
     };
 
     private readonly EventNameManager eventNameManager;
+    private readonly OneCollectorExporterSerializationExceptionStackTraceHandlingType exceptionStackTraceHandling;
 
     public LogRecordCommonSchemaJsonSerializer(
         EventNameManager eventNameManager,
         string tenantToken,
+        OneCollectorExporterSerializationExceptionStackTraceHandlingType exceptionStackTraceHandling = OneCollectorExporterSerializationExceptionStackTraceHandlingType.Ignore,
         int maxPayloadSizeInBytes = int.MaxValue,
         int maxNumberOfItemsPerPayload = int.MaxValue)
         : base(tenantToken, maxPayloadSizeInBytes, maxNumberOfItemsPerPayload)
@@ -68,32 +86,39 @@ internal sealed class LogRecordCommonSchemaJsonSerializer : CommonSchemaJsonSeri
         Debug.Assert(eventNameManager != null, "eventNameManager was null");
 
         this.eventNameManager = eventNameManager!;
+        this.exceptionStackTraceHandling = exceptionStackTraceHandling;
     }
 
     public override string Description => "LogRecord Common Schema JSON";
 
-    protected override void SerializeItemToJson(Resource resource, LogRecord item, Utf8JsonWriter writer)
+    protected override void SerializeItemToJson(Resource resource, LogRecord item, CommonSchemaJsonSerializationState serializationState)
     {
+        Debug.Assert(serializationState != null, "serializationState was null");
+
+        var writer = serializationState!.Writer;
+
+        Debug.Assert(writer != null, "writer was null");
+
         var eventName = this.eventNameManager.ResolveEventFullName(
             item.CategoryName,
             item.EventId.Name);
 
-        writer.WriteStartObject();
+        writer!.WriteStartObject();
 
-        writer.WriteString(VersionProperty, Version4Value);
+        writer.WriteString(CommonSchemaJsonSerializationHelper.VersionProperty, CommonSchemaJsonSerializationHelper.Version4Value);
 
-        writer.WritePropertyName(NameProperty);
+        writer.WritePropertyName(CommonSchemaJsonSerializationHelper.NameProperty);
 #if DEBUG
         writer.WriteRawValue(eventName, skipInputValidation: false);
 #else
         writer.WriteRawValue(eventName, skipInputValidation: true);
 #endif
 
-        writer.WriteString(TimeProperty, item.Timestamp);
+        writer.WriteString(CommonSchemaJsonSerializationHelper.TimeProperty, item.Timestamp);
 
-        writer.WriteString(IKeyProperty, this.TenantTokenWithTenancySystemSymbol);
+        writer.WriteString(CommonSchemaJsonSerializationHelper.IKeyProperty, this.TenantTokenWithTenancySystemSymbol);
 
-        writer.WriteStartObject(DataProperty);
+        writer.WriteStartObject(CommonSchemaJsonSerializationHelper.DataProperty);
 
         /* TODO: There doesn't seem to be a spot in common schema defined for
         event.id so we will drop for now.
@@ -126,6 +151,12 @@ internal sealed class LogRecordCommonSchemaJsonSerializer : CommonSchemaJsonSeri
                     continue;
                 }
 
+                if (AttributeKeyStartWithExtensionPrefix(attribute.Key))
+                {
+                    serializationState.AddExtensionAttribute(attribute);
+                    continue;
+                }
+
                 CommonSchemaJsonSerializationHelper.SerializeKeyValueToJson(attribute.Key, attribute.Value, writer);
             }
         }
@@ -139,26 +170,53 @@ internal sealed class LogRecordCommonSchemaJsonSerializer : CommonSchemaJsonSeri
             writer.WriteString(BodyProperty, item.FormattedMessage);
         }
 
-        if (resource.Attributes is IReadOnlyList<KeyValuePair<string, object?>> resourceAttributeList)
-        {
-            for (int i = 0; i < resourceAttributeList.Count; i++)
-            {
-                var resourceAttribute = resourceAttributeList[i];
+        SerializeResourceToJsonInsideCurrentObject(resource, serializationState);
 
-                CommonSchemaJsonSerializationHelper.SerializeKeyValueToJson(resourceAttribute.Key, resourceAttribute.Value, writer);
-            }
-        }
-        else
-        {
-            foreach (KeyValuePair<string, object> resourceAttribute in resource.Attributes)
-            {
-                CommonSchemaJsonSerializationHelper.SerializeKeyValueToJson(resourceAttribute.Key, resourceAttribute.Value, writer);
-            }
-        }
-
-        item.ForEachScope(SerializeScopeItemToJson, writer);
+        item.ForEachScope(SerializeScopeItemToJson, serializationState);
 
         writer.WriteEndObject();
+
+        this.SerializeExtensionPropertiesToJson(item, writer, serializationState);
+
+        writer.WriteEndObject();
+    }
+
+    private void SerializeExtensionPropertiesToJson(LogRecord item, Utf8JsonWriter writer, CommonSchemaJsonSerializationState serializationState)
+    {
+        var hasTraceContext = item.TraceId != default;
+        var hasException = item.Exception != null;
+
+        if (!hasTraceContext && !hasException && serializationState.ExtensionAttributeCount <= 0)
+        {
+            return;
+        }
+
+        writer.WriteStartObject(CommonSchemaJsonSerializationHelper.ExtensionsProperty);
+
+        if (hasTraceContext)
+        {
+            writer.WriteStartObject(DistributedTraceExtensionProperty);
+            writer.WriteString(DistributedTraceExtensionTraceIdProperty, item.TraceId.ToHexString());
+            writer.WriteString(DistributedTraceExtensionSpanIdProperty, item.SpanId.ToHexString());
+            writer.WriteNumber(DistributedTraceExtensionTraceFlagsProperty, (int)item.TraceFlags);
+            writer.WriteEndObject();
+        }
+
+        if (hasException)
+        {
+            writer.WriteStartObject(ExceptionExtensionProperty);
+            writer.WriteString(ExceptionExtensionTypeProperty, item.Exception!.GetType().FullName);
+            writer.WriteString(ExceptionExtensionMessageProperty, item.Exception.Message);
+
+            if (this.exceptionStackTraceHandling == OneCollectorExporterSerializationExceptionStackTraceHandlingType.IncludeAsString)
+            {
+                writer.WriteString(ExceptionExtensionStackTraceProperty, OneCollectorExporterEventSource.ExceptionToInvariantString(item.Exception));
+            }
+
+            writer.WriteEndObject();
+        }
+
+        serializationState.SerializeExtensionPropertiesToJson(writeExtensionObjectEnvelope: false);
 
         writer.WriteEndObject();
     }